refactor(compaction): remove duplicate position delete collector
related to #1682
