+++ table/compaction/pos_delete_collect.go
@@ -21,7 +21,6 @@ import (
 	"context"
 
 	"github.com/apache/iceberg-go"
-	"github.com/apache/iceberg-go/internal"
 	iceio "github.com/apache/iceberg-go/io"
 	"github.com/apache/iceberg-go/table"
 )
@@ -67,123 +66,4 @@ func CollectDeadPositionDeletes(
 	rewrittenPaths map[string]struct{},
 ) ([]iceberg.DataFile, error) {
 	return table.CollectDeadPositionDeletes(ctx, fs, snap, rewrittenPaths)
-}
-
-type positionDeleteCandidate struct {
-	df               iceberg.DataFile
-	fileScopedTarget string
-	partitionKey     string
-	seq              int64
-}
-
-// collectPositionDeleteCandidates walks the delete manifests and returns the
-// classic position-delete files (deletion vectors excluded), deduplicated by
-// path and resolved to their expunge scope. hasPartitionScoped reports whether
-// any candidate needs the data-manifest survivor pass.
-func collectPositionDeleteCandidates(ctx context.Context, fs iceio.IO, manifests []iceberg.ManifestFile) (candidates []positionDeleteCandidate, hasPartitionScoped bool, err error) {
-	seen := make(map[string]struct{})
-	for _, m := range manifests {
-		if cerr := ctx.Err(); cerr != nil {
-			return nil, false, cerr
-		}
-		if m.ManifestContent() != iceberg.ManifestContentDeletes {
-			continue
-		}
-		for e, err := range m.Entries(fs, true) {
-			if err != nil {
-				return nil, false, err
-			}
-			df := e.DataFile()
-			if df.ContentType() != iceberg.EntryContentPosDeletes || table.IsDeletionVector(df) {
-				continue
-			}
-			path := df.FilePath()
-			if _, ok := seen[path]; ok {
-				continue
-			}
-			seen[path] = struct{}{}
-
-			if target := referencedDataFilePath(df); target != "" {
-				candidates = append(candidates, positionDeleteCandidate{df: df, fileScopedTarget: target})
-
-				continue
-			}
-
-			// Partition-scoped. A negative sequence number is the "unset"
-			// sentinel; retain rather than risk expunging a delete whose
-			// applicability cannot be established.
-			if seq := e.SequenceNum(); seq >= 0 {
-				hasPartitionScoped = true
-				candidates = append(candidates, positionDeleteCandidate{
-					df:           df,
-					partitionKey: partitionBucketKey(df.SpecID(), internal.BorrowedDataFilePartition(df)),
-					seq:          seq,
-				})
-			}
-		}
-	}
-
-	return candidates, hasPartitionScoped, nil
-}
-
-// minSurvivorSeqByPartition walks the data manifests and returns, per (specID,
-// partition), the smallest sequence number among surviving (non-rewritten) data
-// files. A missing key means the partition has no survivor.
-func minSurvivorSeqByPartition(ctx context.Context, fs iceio.IO, manifests []iceberg.ManifestFile, rewrittenPaths map[string]struct{}) (map[string]int64, error) {
-	minSeq := make(map[string]int64)
-	for _, m := range manifests {
-		if cerr := ctx.Err(); cerr != nil {
-			return nil, cerr
-		}
-		if m.ManifestContent() != iceberg.ManifestContentData {
-			continue
-		}
-		for e, err := range m.Entries(fs, true) {
-			if err != nil {
-				return nil, err
-			}
-			df := e.DataFile()
-			if df.ContentType() != iceberg.EntryContentData {
-				continue
-			}
-			if _, rewritten := rewrittenPaths[df.FilePath()]; rewritten {
-				continue
-			}
-			// The negative "unset" sentinel clamps to 0 here — the opposite
-			// bias from the candidate side, which drops unset deletes: an
-			// unknown survivor is treated as old as possible so it retains
-			// every delete it might predate.
-			seq := max(e.SequenceNum(), 0)
-			key := partitionBucketKey(df.SpecID(), internal.BorrowedDataFilePartition(df))
-			if cur, ok := minSeq[key]; ok {
-				seq = min(seq, cur)
-			}
-			minSeq[key] = seq
-		}
-	}
-
-	return minSeq, nil
-}
-
-// decideDeadPositionDeletes returns the candidates that are dead after the
-// rewrite. A file-scoped delete is dead when its single target is rewritten. A
-// partition-scoped delete is dead when no surviving data file in its partition
-// predates it — position deletes apply when data.seq <= delete.seq, so a
-// survivor with seq <= the delete's seq keeps it alive.
-func decideDeadPositionDeletes(candidates []positionDeleteCandidate, rewrittenPaths map[string]struct{}, minSurvivorSeq map[string]int64) []iceberg.DataFile {
-	dead := make([]iceberg.DataFile, 0, len(candidates))
-	for _, c := range candidates {
-		if c.fileScopedTarget != "" {
-			if _, rewritten := rewrittenPaths[c.fileScopedTarget]; rewritten {
-				dead = append(dead, c.df)
-			}
-
-			continue
-		}
-		if cur, ok := minSurvivorSeq[c.partitionKey]; !ok || cur > c.seq {
-			dead = append(dead, c.df)
-		}
-	}
-
-	return dead
 }
